Change LL removeFromTail() to return value of removed node

+++ Program1_ReversePolishCalculator/LinkedList.h
@@ -19,7 +19,7 @@ public:
 	~LinkedList();
 
 	int addToTail(double);
-	int removeFromTail();
+	double removeFromTail();
 };
